PR comments and tests fixes
