+++ internal/cli/atlas/clusters/onlinearchive/create_test.go
@@ -1,0 +1,69 @@
+// Copyright 2020 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// +build unit
+
+package onlinearchive
+
+import (
+	"testing"
+
+	"github.com/golang/mock/gomock"
+	"github.com/mongodb/mongocli/internal/flag"
+	"github.com/mongodb/mongocli/internal/mocks"
+	"github.com/mongodb/mongocli/internal/test"
+	"go.mongodb.org/atlas/mongodbatlas"
+)
+
+func TestCreateBuilder(t *testing.T) {
+	test.CmdValidator(
+		t,
+		CreateBuilder(),
+		0,
+		[]string{
+			flag.ClusterName,
+			flag.Database,
+			flag.Collection,
+			flag.DateField,
+			flag.DateFormat,
+			flag.ArchiveAfter,
+			flag.Partition,
+			flag.ProjectID,
+			flag.Output,
+		},
+	)
+}
+
+func TestCreateOpts_Run(t *testing.T) {
+	ctrl := gomock.NewController(t)
+	mockStore := mocks.NewMockOnlineArchiveCreator(ctrl)
+	defer ctrl.Finish()
+
+	opts := &CreateOpts{
+		store: mockStore,
+	}
+
+	request := opts.newOnlineArchive()
+
+	expected := &mongodbatlas.OnlineArchive{}
+	mockStore.
+		EXPECT().
+		CreateOnlineArchive(opts.ProjectID, opts.clusterName, request).
+		Return(expected, nil).
+		Times(1)
+
+	if err := opts.Run(); err != nil {
+		t.Fatalf("Run() unexpected error: %v", err)
+	}
+}
