+++ internal/cli/atlas/clusters/onlinearchive/watch.go
@@ -1,0 +1,84 @@
+// Copyright 2021 MongoDB Inc
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package onlinearchive
+
+import (
+	"github.com/mongodb/mongocli/internal/cli"
+	"github.com/mongodb/mongocli/internal/cli/require"
+	"github.com/mongodb/mongocli/internal/config"
+	"github.com/mongodb/mongocli/internal/flag"
+	"github.com/mongodb/mongocli/internal/store"
+	"github.com/mongodb/mongocli/internal/usage"
+	"github.com/spf13/cobra"
+)
+
+type WatchOpts struct {
+	cli.GlobalOpts
+	cli.WatchOpts
+	id          string
+	clusterName string
+	store       store.OnlineArchiveDescriber
+}
+
+func (opts *WatchOpts) initStore() error {
+	var err error
+	opts.store, err = store.New(store.PublicAuthenticatedPreset(config.Default()))
+	return err
+}
+
+func (opts *WatchOpts) watcher() (bool, error) {
+	result, err := opts.store.OnlineArchive(opts.ConfigProjectID(), opts.clusterName, opts.id)
+	if err != nil {
+		return false, err
+	}
+	return result.State != "PENDING" && result.State != "PAUSING", nil
+}
+
+func (opts *WatchOpts) Run() error {
+	if err := opts.Watch(opts.watcher); err != nil {
+		return err
+	}
+
+	return opts.Print(nil)
+}
+
+// WatchBuilder
+//	mongocli atlas cluster(s) onlineArchive watch <id>
+//	--clusterName=<name>
+//	[--projectId projectId].
+func WatchBuilder() *cobra.Command {
+	opts := &WatchOpts{}
+	cmd := &cobra.Command{
+		Use:   "watch <id>",
+		Short: "Watch for an archive to be available.",
+		Args:  require.ExactArgs(1),
+		PreRunE: func(cmd *cobra.Command, args []string) error {
+			return opts.PreRunE(
+				opts.ValidateProjectID,
+				opts.initStore,
+				opts.InitOutput(cmd.OutOrStdout(), "\nOnline archive available.\n"),
+			)
+		},
+		RunE: func(cmd *cobra.Command, args []string) error {
+			opts.id = args[0]
+			return opts.Run()
+		},
+	}
+	cmd.Flags().StringVar(&opts.clusterName, flag.ClusterName, "", usage.ClusterName)
+
+	cmd.Flags().StringVar(&opts.ProjectID, flag.ProjectID, "", usage.ProjectID)
+
+	return cmd
+}
